Refactor file and folder structure
