+++ src/pages/configurePatch/configurePatchCore/ConfigureBuildVariants/types.ts
@@ -1,0 +1,7 @@
+interface MenuItemProps {
+  displayName: string;
+  name: string;
+  taskCount: number;
+}
+
+export type { MenuItemProps };
