+++ src/pages/configurePatch/configurePatchCore/ConfigureTasks/types.ts
@@ -1,0 +1,7 @@
+enum CheckboxState {
+  CHECKED = "CHECKED",
+  INDETERMINITE = "INDETERMINITE",
+  UNCHECKED = "UNCHECKED",
+}
+
+export { CheckboxState };
